Align with paper
A set of changes designed to follow the paper of Lee et al (1990) more closely. I define a new set of basis vectors w_i which are transformed back into the original data space, and then attempt to use these rather than the z_i vectors to de-noise the data.